refactor: enable clippy

+++ src/titles/title.rs
@@ -1,14 +1,13 @@
 use crate::survivals::survival_manager::SURVIVAL_MANAGER;
 use crate::swr::*;
 use crate::union_cast;
-use winapi::shared::minwindef::DWORD;
 use winapi::shared::minwindef::LPVOID;
 
 unsafe fn c_title_create(this: LPVOID) -> LPVOID {
     union_cast!(extern "thiscall" fn(this: LPVOID) -> LPVOID)(ORIGINAL_C_TITLE_CREATE)(this)
 }
 
-static mut ORIGINAL_C_TITLE_CREATE: DWORD = 0;
+static mut ORIGINAL_C_TITLE_CREATE: usize = 0;
 
 extern "thiscall" fn c_title_on_create(this: LPVOID) -> LPVOID {
     unsafe {
@@ -20,5 +19,5 @@ extern "thiscall" fn c_title_on_create(this: LPVOID) -> LPVOID {
 }
 
 pub unsafe fn tamper_text() {
-    ORIGINAL_C_TITLE_CREATE = TamperNearJmpOpr(CTitle_Creater, c_title_on_create as DWORD);
+    ORIGINAL_C_TITLE_CREATE = TamperNearJmpOpr(CTitle_Creater, c_title_on_create as usize);
 }
